Integration Domains: add tests to make sure Domains API works
This commit implements unit tests to make sure the app Domain model,
adapter, and view works well. Initially using factory-boy and Faker
libraries for creating the test data.

Coverage: test domain __str__() method and adapter class so that the values for
each of the fields in the serializer match the values for each of the fields on
the model. Views unit tests are also added to make sure the domains API endpoints
(GET, POST, PUT, and PATCH) work as expected.

Signed-off-by: Francisco Trivino <ftrivino@redhat.com>